Semester_5/ProgramAndSoftwareInformationSecurity/лр4/fullstack-auth-broiler/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic-settings==2.2.1
passlib[argon2]==1.7.4
structlog==24.2.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9

pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0